+++ lib/gorge/models/round.rb
@@ -2,8 +2,9 @@
 
 module Gorge
   class Round < Sequel::Model
+    many_to_one :map
+    one_to_many :player_rounds
     many_to_one :server
     many_to_one :winning_team, class: :'Gorge::Team'
-    one_to_many :player_rounds
   end
 end